'second'

+++ app/Http/Controllers/Api/CategoryController.php
@@ -3,23 +3,41 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\StoreCategoryRequest;
-use App\Http\Requests\UpdateCategoryRequest;
+use App\Http\Requests\CategoryRequest;
 use App\Http\Resources\CategoryResource;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use Spatie\FlareClient\Http\Response;
+use function MongoDB\BSON\toJSON;
 
 class CategoryController extends Controller
 {
     /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Resources\Json\AnonymousResourceCollection|\Illuminate\Http\Response
+     * @SWG\Get(
+     *     path="/",
+     *     summary="Возвращает список категорий",
+     *     tags={"categories"},
+     *     @SWG\Response(
+     *         response=200,
+     *         description="успешная операция",
+     *         @SWG\Schema(
+     *             type="array",
+     *             @SWG\Items(ref="#/definitions/Category")
+     *         ),
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Доступ запрещен",
+     *     ),
+     * )
      */
     public function index()
     {
-        return CategoryResource::collection(Category::paginate(5));
+        if (auth()->user()->role_id == 1) {
+            return CategoryResource::collection(Category::paginate(5));
+        }
+        return response('Недостаточно прав', 403);
     }
 
     /**
@@ -27,41 +45,73 @@ class CategoryController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create(StoreCategoryRequest $request)
+    public function create(CategoryRequest $request)
     {
 
     }
 
     /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\Routing\ResponseFactory|\Illuminate\Http\Response
+     * @SWG\Post(
+     *     path="/",
+     *     summary="Создает категорию",
+     *     tags={"categories"},
+     *     @SWG\Response(
+     *         response=200,
+     *         description="Категория создана",
+     *         @SWG\Schema(
+     *             type="category",
+     *             @SWG\Items(ref="#/definitions/Category")
+     *         ),
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Доступ запрещен",
+     *     ),
+     * )
      */
-    public function store(StoreCategoryRequest $request)
+    public function store(CategoryRequest $request)
     {
-        $request->validate([
-            'name' => 'required', 'min:3',
-            'image' => 'required','image','mimes:jpeg,bmp,png'
-        ]);
+        if (auth()->user()->role_id == 1) {
+            $request->validate([
+                'name' => 'required', 'min:3',
+                'image' => 'required','image','mimes:jpeg,bmp,png'
+            ]);
 
-        $category = Category::create([
-            "name" => $request->get("name"),
-            "image" => $request->file('image')->store('public/images')
-        ]);
+            $category = Category::create([
+                "name" => $request->get("name"),
+                "image" => $request->file('image')->store('public/images')
+            ]);
 
             return Response('created',200);
         }
+        return response('Недостаточно прав', 403);
+        }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Category  $category
-     * @return CategoryResource
+     * @SWG\Get(
+     *     path="/{id}",
+     *     summary="Возвращает выбранную категорию",
+     *     tags={"categories"},
+     *     @SWG\Response(
+     *         response=200,
+     *         description="выведена категория",
+     *         @SWG\Schema(
+     *             type="array",
+     *             @SWG\Items(ref="#/definitions/Category")
+     *         ),
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Доступ запрещен",
+     *     ),
+     * )
      */
     public function show(Category $category)
     {
-        return new CategoryResource($category);
+        if (auth()->user()->role_id == 1) {
+            return new CategoryResource($category);
+        }
+        return response('Недостаточно прав', 403);
     }
 
     /**
@@ -84,27 +134,76 @@ class CategoryController extends Controller
      * @return \Illuminate\Http\Response
      */
 
-    public function update(Category $category, UpdateCategoryRequest $request)
+    /**
+     * @SWG\Put(
+     *     path="/{category}",
+     *     summary="Обновляет категорию по id",
+     *     tags={"categories"},
+     *     @SWG\Response(
+     *         response=200,
+     *         description="Категория обновлена",
+     *         @SWG\Schema(
+     *             type="array",
+     *             @SWG\Items(ref="#/definitions/Category")
+     *         ),
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Доступ запрещен",
+     *     ),
+     * )
+     */
+
+    public function update(Request $request, Category $category)
     {
-        dd("test");
+        if (auth()->user()->role_id == 1) {
+            if ($request->hasFile('image')) {
+                Storage::delete("$category->image");
+                $imageName = $request->file('image')->store('public/images');
+            } else {
+                $imageName = $category->image;
+            }
 
-        $category->update([
-            'name' => $request->name,
-            'image' => $request->file('image')->store('public/images')
-        ]);
+            $categoryData = [
+                'name' => $request->name,
+                'image' => $imageName
+            ];
 
-        return Response('updated',200);
+            $category->update($categoryData);
+
+
+            return Response('updated', 200);
+        }
+        return Response('Недостаточно прав', 403);
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Category  $category
-     * @return \Illuminate\Http\Response
+     * @SWG\Get(
+     *     path="/{id}",
+     *     summary="Удаление категории по id",
+     *     tags={"categories"},
+     *     @SWG\Response(
+     *         response=200,
+     *         description="Категория удалена",
+     *         @SWG\Schema(
+     *             type="array",
+     *             @SWG\Items(ref="#/definitions/Category")
+     *         ),
+     *     ),
+     *     @SWG\Response(
+     *         response="403",
+     *         description="Доступ запрещен",
+     *     ),
+     * )
      */
+
     public function destroy(Category $category)
     {
+        if (auth()->user()->role_id == 1) {
+        Storage::delete("$category->image");
         $category->delete();
         return response('deleted',200);
+        }
+        return response('Недостаточно прав', 403);
     }
 }
